+++ app/src/main/java/com/example/tasksapp/ui/TaskViewModel.kt
@@ -1,18 +1,51 @@
 package com.example.tasksapp.ui
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.example.tasksapp.data.model.Status
 import com.example.tasksapp.util.FirebaseHelper
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.ValueEventListener
 import com.myo.tasksapp.data.model.Task
 
 class TaskViewModel : ViewModel() {
+
+    private val _taskList = MutableLiveData<List<Task>>()
+    val taskList: LiveData<List<Task>> = _taskList
 
     private val _taskInsert = MutableLiveData<Task>()
     val taskInsert: LiveData<Task> = _taskInsert
 
     private val _taskUpdate = MutableLiveData<Task>()
     val taskUpdate: LiveData<Task> = _taskUpdate
+
+    fun getTasks(status: Status) {
+        FirebaseHelper.getDatabase()
+            .child("tasks")
+            .child(FirebaseHelper.getUserID())
+            .addListenerForSingleValueEvent(object : ValueEventListener {
+                override fun onDataChange(snapshot: DataSnapshot) {
+
+                    val taskList = mutableListOf<Task>()
+
+                    for (ds in snapshot.children) {
+                        val task = ds.getValue(Task::class.java) as Task
+                        if (task.status == status) {
+                            taskList.add(task)
+                        }
+                    }
+                    taskList.reverse()
+                    _taskList.postValue(taskList)
+                }
+
+                override fun onCancelled(error: DatabaseError) {
+                    Log.i("INFOTESTE", "onCancelled:")
+                }
+            } )
+    }
 
     fun insertTask(task: Task) {
         FirebaseHelper.getDatabase()
