+++ app/src/main/java/com/example/tasksapp/ui/TodoFragment.kt
@@ -1,7 +1,6 @@
 package com.example.tasksapp.ui
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -18,19 +17,14 @@ import com.example.tasksapp.data.model.Status
 import com.example.tasksapp.databinding.FragmentTodoBinding
 import com.example.tasksapp.util.FirebaseHelper
 import com.example.tasksapp.util.showBottomSheet
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.auth
-import com.google.firebase.database.DataSnapshot
-import com.google.firebase.database.DatabaseError
-import com.google.firebase.database.DatabaseReference
-import com.google.firebase.database.ValueEventListener
-import com.google.firebase.database.ktx.database
-import com.google.firebase.ktx.Firebase
 import com.myo.tasksapp.data.model.Task
 import com.myo.tasksapp.ui.adapter.TaskAdapter
 import com.myo.tasksapp.ui.adapter.TaskTopAdapter
 
 
+@Suppress("UnusedEquals", "UnusedEquals", "UnusedEquals", "UnusedEquals", "UnusedEquals",
+    "UnusedEquals"
+)
 class TodoFragment : Fragment() {
 
     private var _binding: FragmentTodoBinding? = null
@@ -53,7 +47,8 @@ class TodoFragment : Fragment() {
 
         initListeners()
         initRecyclerView()
-        getTasks()
+        observerViewModel()
+        viewModel.getTasks(Status.TODO)
     }
 
     private fun initListeners() {
@@ -66,6 +61,13 @@ class TodoFragment : Fragment() {
     }
 
     private fun observerViewModel() {
+
+        viewModel.taskList.observe(viewLifecycleOwner) { taskList ->
+            binding.todoFragmentPB.isVisible = false
+            listEmpty(taskList)
+            taskAdapter.submitList(taskList)
+        }
+
         viewModel.taskInsert.observe(viewLifecycleOwner) { task ->
             if (task.status == Status.TODO) {
 
@@ -179,32 +181,7 @@ class TodoFragment : Fragment() {
         }
     }
 
-    private fun getTasks() {
-        FirebaseHelper.getDatabase()
-            .child("tasks")
-            .child(FirebaseHelper.getUserID())
-            .addListenerForSingleValueEvent(object : ValueEventListener {
-                override fun onDataChange(snapshot: DataSnapshot) {
 
-                    val taskList = mutableListOf<Task>()
-
-                    for (ds in snapshot.children) {
-                        val task = ds.getValue(Task::class.java) as Task
-                        if (task.status == Status.TODO) {
-                            taskList.add(task)
-                        }
-                    }
-                    binding.todoFragmentPB.isVisible = false
-                    listEmpty(taskList)
-                    taskList.reverse()
-                    taskAdapter.submitList(taskList)
-                }
-
-                override fun onCancelled(error: DatabaseError) {
-                    Log.i("INFOTESTE", "onCancelled:")
-                }
-            } )
-    }
 
     private fun deleteTask(task: Task) {
         FirebaseHelper.getDatabase()
